requirements.txt
requests>=2.31.0
anthropic>=0.18.0
python-dotenv>=1.0.0
